Alik

+++ client/src/App.tsx
@@ -7,7 +7,6 @@ import Login, {loader as loginLoader} from "@/Login/Login.tsx";
 import Explore, {loader as exploreLoader} from "@/Main/explore/Explore.tsx";
 import Wallet, {loader as walletLoader} from "@/Main/wallet/Wallet.tsx"
 import Settings, {loader as settingsLoader} from "@/Main/Settings.tsx";
-import axios from "axios";
 
 const router = createBrowserRouter([
     {
@@ -19,12 +18,21 @@ const router = createBrowserRouter([
         path: "/logout",
         async loader() {
             try {
-                await axios.post('http://localhost:3000/api/auth/logout', {}, { withCredentials: true });
+                const response = await fetch('http://localhost:3000/api/auth/logout', {
+                    method: 'POST',
+                    credentials: 'include',
+                    headers: {
+                        'Content-Type': 'application/json'
+                    },
+                    body: JSON.stringify({})
+                });
+                if (!response.ok) {
+                    return new Error('Logout failed');
+                }
                 localStorage.clear();
-                return replace('/login'); // This will handle the redirection
+                return replace('/login');
             } catch (err) {
                 console.error('Logout failed', err);
-                // You may want to handle error here, like returning a response with an error status
                 throw new Error('Logout failed');
             }
         }
@@ -68,7 +76,7 @@ const router = createBrowserRouter([
 
 function App() {
     return (
-        <ThemeProvider defaultTheme='dark' storageKey="vite-ui-theme">
+        <ThemeProvider defaultTheme='system' storageKey="vite-ui-theme">
             <RouterProvider router={router}/>
         </ThemeProvider>
     )
